main-project/legacy_bestbuy/requirements_mac.txt
numpy==1.22.4
openai==1.37.0
pandas==1.3.1
matplotlib==3.4.2
pillow==10.4.0
plotly==5.23.0
pytest==8.3.3
python-dotenv==1.0.1
scikit-learn==0.24.2
seaborn==0.13.2
sentence-transformers==3.0.1
torch==2.0.0
transformers==4.43.2
requests==2.26.0
tensorflow-macos==2.5.0
tensorflow-metal==0.1.2
